Make simple-list into a package and move tests into a separate file.
