+++ examples/simple-lisp/package.json
@@ -1,0 +1,15 @@
+{
+  "private": true,
+  "dependencies": {
+    "ohm-js": "^15.4.1"
+  },
+  "devDependencies": {
+    "nodemon": "^2.0.7",
+    "tap-spec": "^5.0.0",
+    "tape": "^5.2.0"
+  },
+  "scripts": {
+    "test": "node src/simple-lisp.test.mjs",
+    "test:watch": "nodemon -e mjs --exec npm test"
+  }
+}
